Move source.py to models.py
