+++ signal-sweep/src/signal_sweep/core/models.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass
 
 from ..shared.constants import SourceType
 
+
 @dataclass
 class Source:
     url: str
