+++ signal-sweep/tests/handlers/test_handle_text.py
@@ -3,7 +3,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 from itertools import product
 
 from src.signal_sweep.core.handlers.text_handler import TextHandler, _parse_text
-from src.signal_sweep.core.source import Source
+from src.signal_sweep.core.models import Source
 from src.signal_sweep.core.signal_stream import StreamData
